+++ src/localization.rs
@@ -195,5 +195,6 @@ mod test {
         channel_beta: String,
         channel_alpha: String,
         catalog_results: String,
+        categories: String,
     }
 }