+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and `Removed`.
 
 ## [Unreleased]
 
+## Added
+
+- Optional 'Categories' column for Catalog.
+
 ## [0.7.0] - 2021-01-26
 
 ### Added